Bug #71310, proxy necessary methods for CoreLifecycleService and VSLifecycleService
